+++ BoilerPlate/asteroide.cpp
@@ -44,9 +44,9 @@ namespace Asteroids
 			ApplyRandomImpulse();
 		}
 
-		void Asteroid::Update(float deltaTime)
+		void Asteroid::Update(double deltaTime)
 		{
-			float angle = m_transforms->GetAngleInDegrees() + ROTATION_SPEED * deltaTime;
+			float angle = m_transforms->GetAngleInDegrees() + ROTATION_SPEED * static_cast<float>(deltaTime);
 			m_transforms->RotateInDegrees(angle);
 
 			Entity::Update(deltaTime);
