+++ BoilerPlate/update.h
@@ -9,7 +9,7 @@ namespace Engine
 			IUpdate() : m_nUpdates(0) {}
 			virtual ~IUpdate() = default;
 
-			void virtual Update(float deltaTime) { m_nUpdates++; };
+			void virtual Update(double deltaTime) { m_nUpdates++; };
 	    	protected:
 			
 			int	m_nUpdates;
